+++ extensions/apidoc/commands/GuideController.php
@@ -27,6 +27,10 @@ class GuideController extends BaseController
      * @var string path or URL to the api docs to allow links to classes and properties/methods.
      */
     public $apiDocs;
+    /**
+     * @var string prefix to prepend to all output file names generated for the guide.
+     */
+    public $guidePrefix = 'guide-';
 
 
     /**
@@ -43,14 +47,27 @@ class GuideController extends BaseController
             return 1;
         }
 
-        $renderer->guideUrl = './';
+        if ($renderer->guideUrl === null) {
+            $renderer->guideUrl = './';
+        }
+        $renderer->guidePrefix = $this->guidePrefix;
 
         // setup reference to apidoc
         if ($this->apiDocs !== null) {
-            $renderer->apiUrl = $this->apiDocs;
-            $renderer->apiContext = $this->loadContext($this->apiDocs);
+            $path = $this->apiDocs;
+            if ($renderer->apiUrl === null) {
+                $renderer->apiUrl = $path;
+            }
+            // use relative paths relative to targetDir
+            if (strncmp($path, '.', 1) === 0) {
+                $renderer->apiContext = $this->loadContext("$targetDir/$path");
+            } else {
+                $renderer->apiContext = $this->loadContext($path);
+            }
         } elseif (file_exists($targetDir . '/cache/apidoc.data')) {
-            $renderer->apiUrl = './';
+            if ($renderer->apiUrl === null) {
+                $renderer->apiUrl = './';
+            }
             $renderer->apiContext = $this->loadContext($targetDir);
         } else {
             $renderer->apiContext = new Context();
@@ -111,6 +128,6 @@ class GuideController extends BaseController
      */
     public function options($actionID)
     {
-        return array_merge(parent::options($actionID), ['apiDocs']);
+        return array_merge(parent::options($actionID), ['apiDocs', 'guidePrefix']);
     }
 }