+++ extensions/debug/views/default/panels/db/detail.php
@@ -14,7 +14,7 @@ use yii\grid\GridView;
 echo GridView::widget([
     'dataProvider' => $dataProvider,
     'id' => 'db-panel-detailed-grid',
-    'options' => ['class' => 'detail-grid-view'],
+    'options' => ['class' => 'detail-grid-view table-responsive'],
     'filterModel' => $searchModel,
     'filterUrl' => $panel->getUrl(),
     'columns' => [