+++ extensions/sphinx/CHANGELOG.md
@@ -6,6 +6,7 @@ Yii Framework 2 sphinx extension Change Log
 
 - Bug #5601: Simple conditions in Query::where() and ActiveQuery::where() did not allow `yii\db\Expression` to be used as the value (cebe, stevekr)
 - Bug #5634: Fixed `yii\sphinx\QueryBuilder` does not support comparison operators (>,<,>= etc) in where specification (klimov-paul)
+- Bug #6164: Added missing support for `yii\db\Exression` to QueryBuilder `LIKE` conditions (cebe)
 - Enh #5223: Query builder now supports selecting sub-queries as columns (qiangxue)